warehouse, tests: devolve `oidc.models`
This turns `oidc.models` into a directory, with `models.base`
for the basic OIDC publisher classes and adjacent modules
for each specific OIDC implementation (currently just
GitHub).

This will make adding additional OIDC publishers easier
and cleaner, as each will live in its own dedicated submodule.
